Move json formatting in info as util SKIPCI
